Pretty major departure from the current state of ``tornado_cors``.
Updates include:

While an ``Origin`` header does not necessarily indicate a CORS request,
all CORS requests will contain an ``Origin`` header.

[Access-Control-Expose-Headers](http://www.w3.org/TR/cors/#access-control-expose-headers-response-header)

This idea was taken from
[django-cors-headers](https://github.com/ottoyiu/django-cors-headers#configuration).
While it isn't implemented exactly the same, the concept is very
similar.

when the setting is present.

Test cases and associated handlers are aimed at one specific use case.

Many of the settings attributes have updated names to more explicitly
link them to the headers they control. The catalyst for this was
supporting both Allow-Headers and Expose-Headers.